requirements.txt
construct==2.10.70
crcmod==1.7
future==1.0.0
iso8601==2.1.0
pyserial==3.5
PyYAML==6.0.2
